make it uninstall 32bit version if it exists before installing 64bit
This doesn't seem to automatically happen otherwise when running with silent switches.

If run without running silent switches it uninstalls, so I guess this will be needed.

I'm not sure if the uninstall has a silent switch or not, and I don't really plan on testing it at this time, so hopefully it works.